// api/src/utils/interfaces/common.ts
export interface NotificationCategories {
  courseUpdates: boolean;
  assignmentReminders: boolean;
  certificates: boolean;
  systemUpdates: boolean;
}

export interface UpdateSettingsDto {
  theme?: string;
  language?: string;
  timezone?: string;
  dateFormat?: string;
  emailNotif?: boolean;
  pushNotif?: boolean;
  smsNotif?: boolean;
  categories?: NotificationCategories;
}

export interface UserSettingsResponse {
  id: string;
  userId: string;
  theme: string;
  language: string;
  timezone: string;
  dateFormat: string;
  emailNotif: boolean;
  pushNotif: boolean;
  smsNotif: boolean;
  categories: NotificationCategories;
  updatedAt: Date;
}